-make upscaling for hd skins work -get rid of decoder aspect handling, enigma2 will do this now for us, this will also fix wrong aspect ratio on DM800/8000 with 4:3 dvds -implement overlay resizing to fit when using advanced screen modes (16:9 pillarbox, always 16:9 ...) -introduced dynamic choose of scaling routine, use smooth for scaling on sd skins to get readable subtitles, use fast resize for hd skins because time gets more critical there
Attention:
this update needs a modified DVDPlayer frontend within e2, so dont use this
update until enigma2 DVDPlayer Plugin has been updated in git,
otherwise enigma2 will not compile, because we needed some api changes !!!!

Remarks:
there are a lot of possible screen modes in e2, only the main modes have
been tested yet. 7025 is nearly untested with this update, so dont expect
too much, but we needed to commit this for easier and further testing.
16:10 screenmodes are not implemented yet, they will be treaten as 16:9 so
the overlays will not fit exactly, I work on it.