move final "in menu" check down.
SPU and highlight handling could alter "in menu" status so
the final check to report the menu status should be one of
the last things to do before checking user input.
Als check simpla status flag before calling the dvdnav functions
to save a bit of time.
Also set menu on status if domain and in_menu instead of or in_menu
think so.