possible improbeent in highlight handling
ddvd_clear_buttons seems to be used to skip the first highlight
event when 'ok' is pressed by user. This seems to work fine but
occasionally this will render a button invisible when going
back to a previous menu.
So skip the highlight event only when the event seems to be the
same as the previous one.