Use hardware/driver for lower fast forward speed.
Introduce FASTFW and TRICKFW. FASTFW uses the decoder hardware/driver
to do the fast forward.  This results in a much smoother fast forward.
For higher speeds the old method of seeking forward in the stream is
used.
For this the ddvd_trickmode enum is converted to a bitmask enum to
simplify mode checks.