Define debug macro
Move all informational printf into a debug statemets and a level.
Based on the printf is called or not.  The debug level can be set
as a linux environment variable: LIBDVD_DEBUG
Default is 1. 0 is no message at all. 3 or higer gives detailed
info.
Also some cosmetic changes (remove german and unnecessary parenthesis)