Fix resuming DVD
Resuming is based on VTS and subsequent CELL changes.
Sometimes the VTS changes occurs before the resume request is
received. In that case waiting for the next VTS can take a while
when entering the menu.
With this patch the resume checks is moved outsie the VTS_CHANGE code
so it can be processed as soon as the a VTS has passed and the resume
request has been received.